+++ source/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Unreleased
 
+* C/C++/Java: escape newline etc. when printing `String` literals [[#449](https://github.com/BNFC/bnfc/issues/449)]
 * Haskell/GADT: generated `ComposOp.hs` no longer needs `mtl` library [[#438](https://github.com/BNFC/bnfc/pull/438)]
 * Ocaml: fixed a crash in printer with unbalanced `}` [[#439](https://github.com/BNFC/bnfc/issues/439)]
 * Ocaml: removed superfluous `let rec` in the printers for token categories